moved activity logic into activity feed and moved files into shared/activity-feed folder
